feat: Press D to delete the current file, custom keymaps #minor
All the hotkeys are now saved in the `codemade.imageviewer_settings.json` file in the user profile folder. New commands can be defined by adding a new KeyMap that specifies a "Command" element.

Valid commands will use a format like `[path_to_application] {0} [flags]` where `{0}` will be replaced with the current file's path. For example, `cmd.exe /C move \"{0}\" c:\\pictures\\holiday2023\\`. Note that the backslashes are required when inside a JSON string.
Setting the `IsDeleteAction` flag to true, will mark the command as a "delete-like" command, so that ImageViewer will forget about this picture for the session and never show it again.